Outsourced commands and selectors to single files. Fixed spelling error in readme.
