+++ cypress/fixtures/selectors/commands/addFirstItemToBasket.json
@@ -1,0 +1,4 @@
+{
+  "itemList": "li[class='item product product-item']",
+  "addToCartButton": "[title='In den Warenkorb']"
+}
