+++ cypress/fixtures/selectors/commands/getShoppingCarItemCount.json
@@ -1,0 +1,3 @@
+{
+  "miniCartItemCount": "[class='counter-number']"
+}
