+++ cypress/fixtures/selectors/commands/removeItemFromShoppingCart.json
@@ -1,0 +1,4 @@
+{
+  "removeShoppingCartItem": "[title='Artikel entfernen']",
+  "buttonConfirmRemoveFromCart": "[class='action-primary action-accept']"
+}
