+++ cypress/fixtures/selectors/commands/setListLimitAndValidate.json
@@ -1,0 +1,4 @@
+{
+  "itemList": "li[class='item product product-item']",
+  "limiter": "#limiter"
+}
