+++ cypress/fixtures/selectors/commands/sortByPrice.json
@@ -1,0 +1,3 @@
+{
+  "sorter": "#sorter"
+}
