+++ cypress/fixtures/selectors/commands/switchSortingDirection.json
@@ -1,0 +1,3 @@
+{
+  "sorterDirectionSwitch": "[data-role='direction-switcher']"
+}
